requirements.txt
blinker==1.8.2
click==8.1.7
flash==1.0.3
Flask==3.0.3
Flask-Cors==4.0.1
importlib_metadata==8.2.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.1
pillow==10.4.0
pytesseract==0.3.10
Werkzeug==3.0.3
zipp==3.19.2
